+++ packages/video_player/video_player/lib/video_player.dart
@@ -464,6 +464,11 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
           .setMixWithOthers(videoPlayerOptions!.mixWithOthers);
     }
 
+    if (videoPlayerOptions?.allowBackgroundPlayback != null) {
+      await _videoPlayerPlatform.setAllowBackgroundPlayback(
+          videoPlayerOptions!.allowBackgroundPlayback);
+    }
+
     _playerId =
         (await _videoPlayerPlatform.createWithOptions(creationOptions)) ??
             kUninitializedPlayerId;