+++ packages/video_player/video_player_android/pigeons/messages.dart
@@ -47,4 +47,5 @@ abstract class AndroidVideoPlayerApi {
   void seekTo(int playerId, int position);
   void pause(int playerId);
   void setMixWithOthers(bool mixWithOthers);
+  void setAllowBackgroundPlayback(bool allowBackgroundPlayback);
 }