+++ packages/video_player/video_player_avfoundation/lib/src/avfoundation_video_player.dart
@@ -180,6 +180,9 @@ class AVFoundationVideoPlayer extends VideoPlayerPlatform {
   }
 
   @override
+  Future<void> setAllowBackgroundPlayback(bool allowBackgroundPlayback) async {}
+
+  @override
   Widget buildView(int playerId) {
     return buildViewWithOptions(
       VideoViewOptions(playerId: playerId),